removed unused headers, fixed many warnings, few random changes

+++ Toshi/Source/Toshi/Core/TSystem.h
@@ -1,15 +1,14 @@
 #pragma once
-#include "Toshi/Typedefs.h"
 
 namespace Toshi
 {
 	class TSystem
 	{
 	public:
-		static uint32_t StringLength(const char* const& str);
-		static uint32_t StringLength(const wchar_t* const& wstr);
+		static size_t StringLength(const char* const& str);
+		static size_t StringLength(const wchar_t* const& wstr);
 		static void* MemCopy(void* dst, const void* const& src, size_t size);
-		static const char* StringUnicodeToChar(char*, wchar_t*, int);
+		static const char* StringUnicodeToChar(char*, wchar_t*, uint32_t);
 	};
 }
